Fix lint errors in abilities tests and purge ability
