+++ tests/phpunit/test-class-ability.php
@@ -7,82 +7,7 @@
 
 namespace WP_Stream;
 
-/**
- * Concrete fake used to exercise the abstract base.
- *
- * Defined at file scope (not inside a method) so PHP can autoload it.
- */
-class Fake_Ability_For_Test extends Ability {
-
-	/**
-	 * Annotations for the fake ability. Toggled by tests.
-	 *
-	 * @var array
-	 */
-	public $annotations = array();
-
-	/**
-	 * Last input received by execute(), for test inspection.
-	 *
-	 * @var array|null
-	 */
-	public $last_input;
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_name() {
-		return 'stream/fake';
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_label() {
-		return 'Fake';
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_description() {
-		return 'Fake ability used in unit tests.';
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_input_schema() {
-		return array(
-			'type'       => 'object',
-			'properties' => array(
-				'foo' => array( 'type' => 'string' ),
-			),
-		);
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_output_schema() {
-		return array( 'type' => 'string' );
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function get_annotations() {
-		return $this->annotations;
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function execute( $input ) {
-		$this->last_input = $input;
-		return 'ok';
-	}
-}
+require_once __DIR__ . '/fake-ability.php';
 
 /**
  * Class - Test_Ability
